+++ AntWars/AI/AIAntBase.cs
@@ -10,7 +10,7 @@ namespace AntWars.AI
 {
     public abstract class AIAntBase : IAIAnt
     {
-        public Ant Ant { get; internal set; }
+        public MovableAnt Ant { get; internal set; }
         internal Config.GameConfig Conf { get; set; }
         public int BoardWidth { get { return Conf.BoardWidth; } }
         public int BoardHeight { get { return Conf.BoardHeight; } }
